Remove the spring() utilit
The springs object should be used instead.
