+++ src/components/AragonApp/AppBar.js
@@ -3,8 +3,7 @@ import PropTypes from 'prop-types'
 import styled, { css } from 'styled-components'
 import { Transition, animated } from 'react-spring'
 import { PublicUrl } from '../../providers/PublicUrl'
-import { noop, unselectable } from '../../utils'
-import { springs } from '../../utils/styles/spring'
+import { noop, springs, unselectable } from '../../utils'
 import { theme } from '../../theme'
 import Text from '../Text/Text'
